Move the operation builder classes to the same package as the operation classes
