+++ core/src/main/java/com/scalar/db/api/Delete.java
@@ -3,8 +3,8 @@ package com.scalar.db.api;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.common.base.MoreObjects;
-import com.scalar.db.api.builder.DeleteBuilder.BuildableFromExisting;
-import com.scalar.db.api.builder.DeleteBuilder.Namespace;
+import com.scalar.db.api.DeleteBuilder.BuildableFromExisting;
+import com.scalar.db.api.DeleteBuilder.Namespace;
 import com.scalar.db.io.Key;
 import javax.annotation.concurrent.NotThreadSafe;
