+++ core/src/main/java/com/scalar/db/api/GetBuilder.java
@@ -1,18 +1,19 @@
-package com.scalar.db.api.builder;
+package com.scalar.db.api;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-import com.scalar.db.api.Delete;
-import com.scalar.db.api.MutationCondition;
-import com.scalar.db.api.builder.OperationBuilder.ClusteringKey;
-import com.scalar.db.api.builder.OperationBuilder.Condition;
-import com.scalar.db.api.builder.OperationBuilder.Consistency;
-import com.scalar.db.api.builder.OperationBuilder.PartitionKeyBuilder;
-import com.scalar.db.api.builder.OperationBuilder.TableBuilder;
+import com.scalar.db.api.OperationBuilder.ClusteringKey;
+import com.scalar.db.api.OperationBuilder.Consistency;
+import com.scalar.db.api.OperationBuilder.PartitionKeyBuilder;
+import com.scalar.db.api.OperationBuilder.Projection;
+import com.scalar.db.api.OperationBuilder.TableBuilder;
 import com.scalar.db.io.Key;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 import javax.annotation.Nullable;
 
-public class DeleteBuilder {
+public class GetBuilder {
 
   public static class Namespace implements OperationBuilder.Namespace<Table> {
     @Override
@@ -47,14 +48,21 @@ public class DeleteBuilder {
     }
   }
 
-  public static class Buildable extends OperationBuilder.Buildable<Delete>
-      implements ClusteringKey<Buildable>, Consistency<Buildable>, Condition<Buildable> {
+  public static class Buildable extends OperationBuilder.Buildable<Get>
+      implements ClusteringKey<Buildable>, Consistency<Buildable>, Projection<Buildable> {
+    final List<String> projections = new ArrayList<>();
     @Nullable Key clusteringKey;
     @Nullable com.scalar.db.api.Consistency consistency;
-    @Nullable MutationCondition condition;
 
     public Buildable(String namespace, String table, Key partitionKey) {
       super(namespace, table, partitionKey);
+    }
+
+    @Override
+    public Buildable projection(String projection) {
+      checkNotNull(projection);
+      projections.add(projection);
+      return this;
     }
 
     @Override
@@ -65,10 +73,24 @@ public class DeleteBuilder {
     }
 
     @Override
-    public Buildable condition(MutationCondition condition) {
-      checkNotNull(condition);
-      this.condition = condition;
+    public Buildable projections(Collection<String> projections) {
+      checkNotNull(projections);
+      this.projections.addAll(projections);
       return this;
+    }
+
+    @Override
+    public Get build() {
+      Get get = new Get(partitionKey, clusteringKey);
+      get.forNamespace(namespaceName).forTable(tableName);
+      if (!projections.isEmpty()) {
+        get.withProjections(projections);
+      }
+      if (consistency != null) {
+        get.withConsistency(consistency);
+      }
+
+      return get;
     }
 
     @Override
@@ -77,38 +99,20 @@ public class DeleteBuilder {
       this.consistency = consistency;
       return this;
     }
-
-    @Override
-    public Delete build() {
-      Delete delete = new Delete(partitionKey, clusteringKey);
-      delete.forNamespace(namespaceName).forTable(tableName);
-      if (condition != null) {
-        delete.withCondition(condition);
-      }
-      if (consistency != null) {
-        delete.withConsistency(consistency);
-      }
-
-      return delete;
-    }
   }
 
   public static class BuildableFromExisting extends Buildable
       implements OperationBuilder.Namespace<BuildableFromExisting>,
           OperationBuilder.Table<BuildableFromExisting>,
           OperationBuilder.PartitionKey<BuildableFromExisting>,
-          OperationBuilder.ClearCondition<BuildableFromExisting>,
+          OperationBuilder.ClearProjections<BuildableFromExisting>,
           OperationBuilder.ClearClusteringKey<BuildableFromExisting> {
 
-    public BuildableFromExisting(Delete delete) {
-      super(
-          delete.forNamespace().orElse(null),
-          delete.forTable().orElse(null),
-          delete.getPartitionKey());
-
-      this.clusteringKey = delete.getClusteringKey().orElse(null);
-      this.consistency = delete.getConsistency();
-      this.condition = delete.getCondition().orElse(null);
+    public BuildableFromExisting(Get get) {
+      super(get.forNamespace().orElse(null), get.forTable().orElse(null), get.getPartitionKey());
+      this.clusteringKey = get.getClusteringKey().orElse(null);
+      this.projections.addAll(get.getProjections());
+      this.consistency = get.getConsistency();
     }
 
     @Override
@@ -145,14 +149,20 @@ public class DeleteBuilder {
     }
 
     @Override
-    public BuildableFromExisting condition(MutationCondition condition) {
-      super.condition(condition);
+    public BuildableFromExisting projection(String projection) {
+      super.projection(projection);
       return this;
     }
 
     @Override
-    public BuildableFromExisting clearCondition() {
-      this.condition = null;
+    public BuildableFromExisting projections(Collection<String> projections) {
+      super.projections(projections);
+      return this;
+    }
+
+    @Override
+    public BuildableFromExisting clearProjections() {
+      this.projections.clear();
       return this;
     }
