+++ core/src/main/java/com/scalar/db/api/Put.java
@@ -5,8 +5,8 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import com.google.common.base.MoreObjects;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
-import com.scalar.db.api.builder.PutBuilder.BuildableFromExisting;
-import com.scalar.db.api.builder.PutBuilder.Namespace;
+import com.scalar.db.api.PutBuilder.BuildableFromExisting;
+import com.scalar.db.api.PutBuilder.Namespace;
 import com.scalar.db.io.BigIntColumn;
 import com.scalar.db.io.BlobColumn;
 import com.scalar.db.io.BooleanColumn;
